requirements.txt
matplotlib
pandas 
pyarrow
wandb
opencv-python
scipy
scikit-learn
scikit-image
einops
timm
torchmetrics
lightning
fvcore
torchinfo
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
opencv-python
scikit-image
Pillow
timm
lightning
einops
fvcore
torchmetrics
torchinfo